[file] fix NullPointerException CopyHelper.java:290
https://play.google.com/apps/publish/?dev_acc=17736375203683117254#ErrorClusterDetailsPlace:p=imoblife.toolbox.full&et=CRASH&lr=LAST_7_DAYS&ecn=java.lang.NullPointerException&tf=CopyHelper.java&tc=com.filemanager.util.CopyHelper&tm=cutFolderAndroid5&nid&an&c&s=reports_last_week_desc

java.lang.RuntimeException: An error occured while executing doInBackground()
	at imoblife.android.os.ModernAsyncTask$3.done(ModernAsyncTask.java:144)
	at java.util.concurrent.FutureTask.finishCompletion(FutureTask.java:355)
	at java.util.concurrent.FutureTask.setException(FutureTask.java:222)
	at java.util.concurrent.FutureTask.run(FutureTask.java:242)
	at java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1112)
	at java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:587)
	at java.lang.Thread.run(Thread.java:818)
Caused by: java.lang.NullPointerException: Attempt to get length of null array
	at com.filemanager.util.CopyHelper.cutFolderAndroid5(CopyHelper.java:290)
	at com.filemanager.util.CopyHelper.cutFolderAndroid5(CopyHelper.java:300)
	at com.filemanager.util.CopyHelper.performCut(CopyHelper.java:328)
	at com.filemanager.util.CopyHelper.access$5(CopyHelper.java:317)
	at com.filemanager.util.CopyHelper$MoveAsync.doInBackground(CopyHelper.java:423)
	at com.filemanager.util.CopyHelper$MoveAsync.doInBackground(CopyHelper.java:1)
	at imoblife.android.os.ModernAsyncTask$2.call(ModernAsyncTask.java:130)
	at java.util.concurrent.FutureTask.run(FutureTask.java:237)
	... 3 more

+++ FileManager/src/com/filemanager/util/CopyHelper.java
@@ -287,7 +287,7 @@ public class CopyHelper {
 
             // list all the directory contents
             String files[] = oldFile.list();
-            if (files.length == 0) {
+            if (files != null &&  files.length == 0) {
                 return FileUtil.deleteFile(oldFile, mContext);
             }
